register api updated
