--- src/core/services/api/(step2)verifymessage.js
@@ -1,12 +1,0 @@
-import http from "../interceptor";
-
-export const VerifyMessage = async (user) => {
-  try {
-    const response = await http.post("/Sign/VerifyMessage", user);
-
-    return response;
-  } catch (error) {
-    console.log("error",(error));
-    return false;
-  }
-};
