--- src/core/services/api/(step3)register.js
@@ -1,12 +1,0 @@
-import http from "../interceptor";
-
-export const Register = async (user) => {
-  try {
-    const response = await http.post("/Sign/Register", user);
-
-    return response;
-  } catch (error) {
-    console.log("error",(error));
-    return false;
-  }
-};
